Refactor login server login into login directory
